inits folium_viz_system 2.0
