--- mescal/visualizations/folium_map/folium_viz_system.py
@@ -1,812 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import Any, Union, Callable, Optional, List
-from dataclasses import dataclass, field
-import pandas as pd
-import folium
-from shapely.geometry import Point, LineString, Polygon, MultiPolygon
-
-from mescal.kpis import KPI, KPICollection
-
-
-@dataclass
-class StyleMapper:
-    """Maps a data column to a visual property using a mapping function."""
-    property: str  # e.g., 'color', 'width', 'opacity', 'height', 'arrow_speed', 'shadow_color'
-    column: str = None  # column to get value from data source
-    mapping: Union[Callable, Any] = None  # mapper function or static value
-    return_type: type = str  # expected return type for validation
-
-
-@dataclass
-class ResolvedStyle:
-    """Container for resolved style properties."""
-    properties: dict = field(default_factory=dict)
-
-    def get(self, property: str, default=None):
-        return self.properties.get(property, default)
-
-    def __getitem__(self, key):
-        return self.properties[key]
-
-    def __setitem__(self, key, value):
-        self.properties[key] = value
-
-    def __contains__(self, key):
-        return key in self.properties
-
-    # Convenience properties for common styles
-    @property
-    def color(self) -> str:
-        return self.get('color', '#000000')
-
-    @property
-    def width(self) -> float:
-        return self.get('width', 1.0)
-
-    @property
-    def opacity(self) -> float:
-        return self.get('opacity', 1.0)
-
-
-class MapDataItem(ABC):
-    """Abstract interface for data items that can be visualized on maps."""
-
-    @abstractmethod
-    def get_geometry(self) -> Any:
-        """Get the geometric representation of the data."""
-        pass
-
-    @abstractmethod
-    def get_tooltip_data(self) -> dict:
-        """Get data for tooltip display."""
-        pass
-
-    @abstractmethod
-    def get_styling_value(self, column: str) -> Any:
-        """Get value for styling from specified column."""
-        pass
-
-    @abstractmethod
-    def get_location(self) -> tuple[float, float]:
-        """Get lat/lon coordinates for point-based objects."""
-        pass
-
-
-class ModelDataItem(MapDataItem):
-    """Map data item for model DataFrame rows."""
-
-    def __init__(self, object_data: pd.Series):
-        self.object_data = object_data
-        self.object_id = object_data.name
-
-    def get_geometry(self) -> Any:
-        return self.object_data.get('geometry')
-
-    def get_tooltip_data(self) -> dict:
-        data = {'ID': self.object_id}
-        for col, value in self.object_data.items():
-            if pd.notna(value):
-                value_str = str(value)
-                if len(value_str) > 50:
-                    value_str = value_str[:47] + "..."
-                data[col] = value_str
-        return data
-
-    def get_styling_value(self, column: str) -> Any:
-        return self.object_data.get(column)
-
-    def get_location(self) -> tuple[float, float]:
-        if 'location' in self.object_data:
-            location = self.object_data['location']
-            if isinstance(location, Point):
-                return location.y, location.x
-
-        geometry = self.get_geometry()
-        if isinstance(geometry, Point):
-            return geometry.y, geometry.x
-        elif isinstance(geometry, (Polygon, MultiPolygon)):
-            point = geometry.representative_point()
-            return point.y, point.x
-        elif isinstance(geometry, LineString):
-            point = geometry.interpolate(0.5, normalized=True)
-            return point.y, point.x
-
-        raise ValueError(f"Cannot determine location for {self.object_id}")
-
-
-class KPIDataItem(MapDataItem):
-    """Map data item for KPI objects - reuses ModelDataItem internally."""
-
-    def __init__(self, kpi: KPI, kpi_collection: KPICollection = None, study_manager=None):
-        self.kpi = kpi
-        self.kpi_collection = kpi_collection
-        self.study_manager = study_manager
-        self._object_info = kpi.get_attributed_object_info_from_model()
-        self._model_item = ModelDataItem(self._object_info)
-
-    def get_geometry(self) -> Any:
-        return self._model_item.get_geometry()
-
-    def get_tooltip_data(self) -> dict:
-        kpi_data = {
-            'KPI': self.kpi.get_kpi_name_with_dataset_name(),
-            'Value': str(self.kpi.quantity),
-        }
-        model_data = self._model_item.get_tooltip_data()
-        return {**kpi_data, **model_data}
-
-    def get_styling_value(self, column: str) -> Any:
-        if column == 'kpi_value':
-            return self.kpi.value
-        return self._model_item.get_styling_value(column)
-
-    def get_location(self) -> tuple[float, float]:
-        if 'projection_point' in self._object_info:
-            point = self._object_info['projection_point']
-            return point.y, point.x
-        return self._model_item.get_location()
-
-
-class StyleResolver:
-    """Resolves styling for map data items using flexible property mappings."""
-
-    def __init__(self, style_mappers: List[StyleMapper]):
-        self.style_mappers = {mapper.property: mapper for mapper in style_mappers}
-
-    def resolve_style(self, data_item: MapDataItem) -> ResolvedStyle:
-        """Resolve styling for a data item."""
-        resolved = ResolvedStyle()
-
-        for prop, mapper in self.style_mappers.items():
-            if mapper.column:
-                value = data_item.get_styling_value(mapper.column)
-            else:
-                value = None
-
-            if callable(mapper.mapping):
-                resolved[prop] = mapper.mapping(value)
-            else:
-                resolved[prop] = mapper.mapping
-
-        return resolved
-
-    @classmethod
-    def create_simple(
-            cls,
-            color: Union[Callable, str] = '#D9D9D9',
-            width: Union[Callable, float] = 3.0,
-            opacity: Union[Callable, float] = 1.0,
-            color_column: str = None,
-            width_column: str = None,
-            opacity_column: str = None,
-            **additional_mappings
-    ) -> 'StyleResolver':
-        """Create a StyleResolver with common mappings."""
-        mappers = [
-            StyleMapper('color', color_column, color, str),
-            StyleMapper('width', width_column, width, float),
-            StyleMapper('opacity', opacity_column, opacity, float),
-        ]
-
-        # Add additional mappings
-        for prop, mapping_info in additional_mappings.items():
-            if isinstance(mapping_info, dict):
-                mappers.append(StyleMapper(
-                    property=prop,
-                    column=mapping_info.get('column'),
-                    mapping=mapping_info.get('mapping'),
-                    return_type=mapping_info.get('return_type', str)
-                ))
-            else:
-                mappers.append(StyleMapper(prop, None, mapping_info, type(mapping_info)))
-
-        return cls(mappers)
-
-
-class TooltipGenerator:
-    """Generates HTML tooltips from map data items."""
-
-    def generate_tooltip(self, data_item: MapDataItem) -> str:
-        """Generate HTML tooltip from data item."""
-        tooltip_data = data_item.get_tooltip_data()
-
-        html = '<table style="border-collapse: collapse;">\n'
-        for key, value in tooltip_data.items():
-            html += f'  <tr><td style="padding: 4px 8px;"><strong>{key}</strong></td>' \
-                    f'<td style="text-align: right; padding: 4px 8px;">{value}</td></tr>\n'
-        html += '</table>'
-        return html
-
-
-class IconGenerator(ABC):
-    """Abstract base for generating folium icons."""
-
-    @abstractmethod
-    def generate_icon(self, data_item: MapDataItem, style: ResolvedStyle) -> folium.DivIcon:
-        """Generate a folium icon for the data item."""
-        pass
-
-
-class FoliumObjectGenerator(ABC):
-    """Abstract base for generating folium objects."""
-
-    def __init__(
-            self,
-            style_resolver: StyleResolver = None,
-            tooltip_generator: TooltipGenerator = None
-    ):
-        self.style_resolver = style_resolver or StyleResolver([])
-        self.tooltip_generator = tooltip_generator or TooltipGenerator()
-
-    @abstractmethod
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        """Generate folium object and add it to the feature group."""
-        pass
-
-
-class AreaGenerator(FoliumObjectGenerator):
-    """Generates folium GeoJson objects for area geometries."""
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        geometry = data_item.get_geometry()
-        if not isinstance(geometry, (Polygon, MultiPolygon)):
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-
-        style_dict = {
-            'fillColor': style.color,
-            'color': style.get('border_color', 'white'),
-            'weight': style.width,
-            'fillOpacity': style.opacity
-        }
-
-        highlight_dict = style_dict.copy()
-        highlight_dict['weight'] = style.width * 1.5
-        highlight_dict['fillOpacity'] = min(style.opacity * 1.5, 1.0)
-
-        geojson_data = {
-            "type": "Feature",
-            "geometry": geometry.__geo_interface__,
-            "properties": {"tooltip": tooltip}
-        }
-
-        folium.GeoJson(
-            geojson_data,
-            style_function=lambda x, s=style_dict: s,
-            highlight_function=lambda x, h=highlight_dict: h,
-            tooltip=folium.GeoJsonTooltip(fields=['tooltip'], aliases=[''], sticky=True)
-        ).add_to(feature_group)
-
-
-class LineGenerator(FoliumObjectGenerator):
-    """Generates folium PolyLine objects for line geometries."""
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        geometry = data_item.get_geometry()
-        if not isinstance(geometry, LineString):
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-
-        coordinates = [(lat, lon) for lon, lat in geometry.coords]
-
-        line_kwargs = {
-            'locations': coordinates,
-            'color': style.color,
-            'weight': style.width,
-            'opacity': style.opacity,
-            'tooltip': tooltip
-        }
-
-        # Add dash pattern if specified
-        if 'dash_pattern' in style:
-            line_kwargs['dashArray'] = style['dash_pattern']
-
-        folium.PolyLine(**line_kwargs).add_to(feature_group)
-
-
-class NodeGenerator(FoliumObjectGenerator):
-    """Generates folium Marker or CircleMarker objects for point geometries."""
-
-    def __init__(
-            self,
-            style_resolver: StyleResolver = None,
-            tooltip_generator: TooltipGenerator = None,
-            icon_generator: IconGenerator = None,
-            use_circle_marker: bool = True
-    ):
-        super().__init__(style_resolver, tooltip_generator)
-        self.icon_generator = icon_generator
-        self.use_circle_marker = use_circle_marker
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        try:
-            location = data_item.get_location()
-        except ValueError:
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        tooltip = self.tooltip_generator.generate_tooltip(data_item)
-
-        if self.icon_generator:
-            icon = self.icon_generator.generate_icon(data_item, style)
-            folium.Marker(
-                location=location,
-                icon=icon,
-                tooltip=tooltip
-            ).add_to(feature_group)
-        elif self.use_circle_marker:
-            folium.CircleMarker(
-                location=location,
-                radius=style.width,
-                color=style.get('border_color', 'white'),
-                fillColor=style.color,
-                fillOpacity=style.opacity,
-                weight=style.get('border_width', 1),
-                tooltip=tooltip
-            ).add_to(feature_group)
-        else:
-            folium.Marker(
-                location=location,
-                tooltip=tooltip
-            ).add_to(feature_group)
-
-
-class TextOverlayGenerator(FoliumObjectGenerator):
-    """Generates text overlays for map data items."""
-
-    def __init__(
-            self,
-            style_resolver: StyleResolver = None,
-            tooltip_generator: TooltipGenerator = None,
-            text_formatter: Callable[[MapDataItem], str] = None
-    ):
-        super().__init__(style_resolver, tooltip_generator)
-        self.text_formatter = text_formatter or self._default_text_formatter
-
-    def _default_text_formatter(self, data_item: MapDataItem) -> str:
-        if isinstance(data_item, KPIDataItem):
-            return f"{data_item.kpi.value:.1f}"
-        return str(data_item.object_id)
-
-    def generate(self, data_item: MapDataItem, feature_group: folium.FeatureGroup) -> None:
-        try:
-            location = data_item.get_location()
-        except ValueError:
-            return
-
-        style = self.style_resolver.resolve_style(data_item)
-        text = self.text_formatter(data_item)
-
-        text_color = style.get('text_color', self._get_contrasting_color(style.color))
-        shadow_color = style.get('shadow_color', self._get_shadow_color(text_color))
-        font_size = style.get('font_size', '10pt')
-
-        icon_html = f'''
-            <div style="
-                position: absolute;
-                left: 50%;
-                top: 50%;
-                transform: translate(-50%, -50%);
-                text-align: center;
-                font-size: {font_size};
-                font-weight: bold;
-                color: {text_color};
-                white-space: nowrap;
-                text-shadow:
-                   -0.5px -0.5px 0 {shadow_color},  
-                    0.5px -0.5px 0 {shadow_color},
-                   -0.5px  0.5px 0 {shadow_color},
-                    0.5px  0.5px 0 {shadow_color};
-            ">
-                {text}
-            </div>
-        '''
-
-        folium.Marker(
-            location=location,
-            icon=folium.DivIcon(html=icon_html)
-        ).add_to(feature_group)
-
-    def _get_contrasting_color(self, surface_color: str) -> str:
-        """Get contrasting text color for a surface color."""
-        if self._is_dark(surface_color):
-            return '#F2F2F2'
-        return '#3A3A3A'
-
-    def _get_shadow_color(self, text_color: str) -> str:
-        """Get shadow color for text."""
-        if text_color == '#F2F2F2':
-            return '#3A3A3A'
-        return '#F2F2F2'
-
-    @staticmethod
-    def _is_dark(color: str) -> bool:
-        """Check if a color is dark."""
-        if not color.startswith('#'):
-            return False
-        try:
-            r, g, b = [int(color[i:i + 2], 16) for i in (1, 3, 5)]
-            return (0.299 * r + 0.587 * g + 0.114 * b) < 160
-        except (ValueError, IndexError):
-            return False
-
-
-class FoliumMapBuilder:
-    """High-level builder for creating folium visualizations."""
-
-    def __init__(self):
-        self.generators: dict[str, FoliumObjectGenerator] = {}
-
-    def register_generator(self, name: str, generator: FoliumObjectGenerator):
-        """Register a generator for a specific visualization type."""
-        self.generators[name] = generator
-
-    def add_model_data(
-            self,
-            generator: Union[str, FoliumObjectGenerator],
-            model_df: pd.DataFrame,
-            feature_group: folium.FeatureGroup
-    ):
-        """Add model DataFrame data to the map."""
-        gen = self._resolve_generator(generator)
-
-        for _, row in model_df.iterrows():
-            data_item = ModelDataItem(row)
-            gen.generate(data_item, feature_group)
-
-    def add_kpi_data(
-            self,
-            generator: Union[str, FoliumObjectGenerator],
-            kpi_collection: KPICollection,
-            feature_group: folium.FeatureGroup,
-            study_manager=None
-    ):
-        """Add KPI data to the map."""
-        gen = self._resolve_generator(generator)
-
-        for kpi in kpi_collection:
-            data_item = KPIDataItem(kpi, kpi_collection, study_manager)
-            gen.generate(data_item, feature_group)
-
-    def add_single_kpi(
-            self,
-            generator: Union[str, FoliumObjectGenerator],
-            kpi: KPI,
-            feature_group: folium.FeatureGroup,
-            kpi_collection: KPICollection = None,
-            study_manager=None
-    ):
-        """Add a single KPI to the map with optional context."""
-        gen = self._resolve_generator(generator)
-        data_item = KPIDataItem(kpi, kpi_collection, study_manager)
-        gen.generate(data_item, feature_group)
-
-    def _resolve_generator(self, generator: Union[str, FoliumObjectGenerator]) -> FoliumObjectGenerator:
-        """Resolve generator from name or return instance directly."""
-        if isinstance(generator, str):
-            if generator not in self.generators:
-                raise ValueError(f"Generator '{generator}' not registered")
-            return self.generators[generator]
-        return generator
-
-
-class KPIGroupingManager:
-    """Handles KPI grouping logic extracted from KPIToMapVisualizerBase."""
-
-    def __init__(
-            self,
-            kpi_attribute_category_orders: dict[str, list[str]] = None,
-            kpi_attribute_keys_to_exclude_from_grouping: list[str] = None,
-            kpi_attribute_sort_order: list[str] = None
-    ):
-        self.kpi_attribute_category_orders = kpi_attribute_category_orders or {}
-        self.kpi_attribute_keys_to_exclude_from_grouping = kpi_attribute_keys_to_exclude_from_grouping or [
-            'name', 'object_name', 'column_subset'
-        ]
-        self.kpi_attribute_sort_order = kpi_attribute_sort_order or [
-            'name_prefix', 'model_flag', 'flag', 'model_query', 'aggregation',
-            'reference_dataset', 'variation_dataset', 'dataset',
-            'value_comparison', 'value_operation', 'name_suffix'
-        ]
-
-    def get_kpi_groups(self, kpi_collection: KPICollection) -> list[KPICollection]:
-        """Group KPIs by attributes with sophisticated sorting."""
-        from mescal.utils.dict_combinations import dict_combination_iterator
-
-        attribute_sets = kpi_collection.get_all_kpi_attributes_and_value_sets(primitive_values=True)
-        relevant_attribute_sets = {
-            k: v for k, v in attribute_sets.items()
-            if k not in self.kpi_attribute_keys_to_exclude_from_grouping
-        }
-
-        ordered_keys = [k for k in self.kpi_attribute_sort_order if k in relevant_attribute_sets]
-
-        # Build attribute value rankings
-        attribute_value_rank: dict[str, dict[str, int]] = {}
-        for attr in ordered_keys:
-            existing_values = set(relevant_attribute_sets.get(attr, []))
-            manual_order = [v for v in self.kpi_attribute_category_orders.get(attr, []) if v in existing_values]
-            remaining = sorted(existing_values - set(manual_order))
-            full_order = manual_order + remaining
-            attribute_value_rank[attr] = {val: idx for idx, val in enumerate(full_order)}
-
-        def sorting_index(group_kwargs: dict[str, str]) -> tuple:
-            return tuple(
-                attribute_value_rank[attr].get(group_kwargs.get(attr), float("inf"))
-                for attr in ordered_keys
-            )
-
-        # Create and sort groups
-        group_kwargs_list = list(dict_combination_iterator(relevant_attribute_sets))
-        group_kwargs_list.sort(key=sorting_index)
-
-        groups: list[KPICollection] = []
-        for group_kwargs in group_kwargs_list:
-            g = kpi_collection.get_filtered_kpi_collection_by_attributes(**group_kwargs)
-            if not g.empty:
-                groups.append(g)
-
-        return groups
-
-    def get_feature_group_name(self, kpi_group: KPICollection) -> str:
-        """Generate meaningful feature group name from KPI group."""
-        _include = ['value_operation', 'aggregation', 'flag', 'dataset', 'unit']
-        _exclude = ['variation_dataset', 'reference_dataset', 'model_flag', 'base_unit', 'dataset_type']
-
-        attributes = kpi_group.get_in_common_kpi_attributes(primitive_values=True)
-        for k in _exclude:
-            attributes.pop(k, None)
-
-        components = []
-        _include += [k for k in attributes.keys() if k not in _include]
-        for k in _include:
-            value = attributes.pop(k, None)
-            if value is not None:
-                components.append(str(value))
-
-        return ' '.join(components)
-
-    def get_related_kpi_groups(self, kpi: KPI, study_manager) -> dict[str, KPICollection]:
-        """Get related KPIs grouped by relationship type."""
-        from mescal.kpis import ValueComparisonKPI, ArithmeticValueOperationKPI
-
-        groups = {
-            'Different Comparisons / ValueOperations': KPICollection(),
-            'Different Aggregations': KPICollection(),
-            'Different Datasets': KPICollection(),
-        }
-
-        if not study_manager:
-            return groups
-
-        kpi_atts = kpi.attributes.as_dict(primitive_values=True)
-
-        _must_contain = ['flag', 'aggregation']
-        if any(kpi_atts.get(k, None) is None for k in _must_contain):
-            return groups
-
-        try:
-            pre_filtered = study_manager.scen_comp.get_merged_kpi_collection()
-            pre_filtered = pre_filtered.get_filtered_kpi_collection_by_attributes(
-                object_name=kpi.get_attributed_object_name(),
-                flag=kpi_atts['flag'],
-                model_flag=kpi.get_attributed_model_flag(),
-            )
-        except:
-            return groups
-
-        _main_kpi_is_value_op = isinstance(kpi, (ValueComparisonKPI, ArithmeticValueOperationKPI))
-
-        for potential_relative in pre_filtered:
-            pratts = potential_relative.attributes.as_dict(primitive_values=True)
-            if pratts.get('dataset') == kpi_atts.get('dataset'):  # same ds
-                if pratts.get('aggregation', None) == kpi_atts.get('aggregation'):  # same ds, agg
-                    if pratts.get('value_operation', None) != kpi_atts.get('value_operation', None):
-                        groups['Different Comparisons / ValueOperations'].add_kpi(potential_relative)
-                        continue
-                else:  # same ds, diff agg
-                    if pratts.get('value_operation', None) is None:
-                        groups['Different Aggregations'].add_kpi(potential_relative)
-                        continue
-                    elif pratts.get('value_operation') == kpi_atts.get('value_operation', None):
-                        groups['Different Aggregations'].add_kpi(potential_relative)
-                        continue
-            elif pratts.get('aggregation', None) == kpi_atts.get('aggregation'):  # same agg, diff ds
-                if pratts.get('value_operation', None) == kpi_atts.get('value_operation', None):
-                    groups['Different Datasets'].add_kpi(potential_relative)
-                    continue
-                if not _main_kpi_is_value_op:
-                    groups['Different Comparisons / ValueOperations'].add_kpi(potential_relative)
-                    continue
-
-        return groups
-
-
-class KPIMapVisualizer:
-    """High-level KPI map visualizer that replicates KPIToMapVisualizerBase functionality."""
-
-    def __init__(
-            self,
-            generator: FoliumObjectGenerator,
-            study_manager=None,
-            include_related_kpis_in_tooltip: bool = False,
-            kpi_attribute_category_orders: dict[str, list[str]] = None,
-            kpi_attribute_keys_to_exclude_from_grouping: list[str] = None,
-            kpi_attribute_sort_order: list[str] = None
-    ):
-        self.generator = generator
-        self.study_manager = study_manager
-        self.include_related_kpis_in_tooltip = include_related_kpis_in_tooltip
-
-        self.grouping_manager = KPIGroupingManager(
-            kpi_attribute_category_orders,
-            kpi_attribute_keys_to_exclude_from_grouping,
-            kpi_attribute_sort_order
-        )
-
-        # Enhanced tooltip if needed
-        if self.include_related_kpis_in_tooltip:
-            self.generator.tooltip_generator = self._create_enhanced_tooltip_generator()
-
-    def get_feature_groups(self, kpi_collection: KPICollection) -> list[folium.FeatureGroup]:
-        """Create feature groups for KPI collection, replicating original functionality."""
-        from tqdm import tqdm
-        from mescal.utils.logging import get_logger
-
-        logger = get_logger(__name__)
-        feature_groups = []
-
-        pbar = tqdm(kpi_collection, total=kpi_collection.size, desc=f'{self.__class__.__name__}')
-        with pbar:
-            for kpi_group in self.grouping_manager.get_kpi_groups(kpi_collection):
-                group_name = self.grouping_manager.get_feature_group_name(kpi_group)
-                fg = folium.FeatureGroup(name=group_name, overlay=False, show=False)
-
-                for kpi in kpi_group:
-                    try:
-                        data_item = KPIDataItem(kpi, kpi_collection, self.study_manager)
-                        self.generator.generate(data_item, fg)
-                    except Exception as e:
-                        logger.warning(
-                            f'Exception while trying to add KPI {kpi.name} to FeatureGroup {group_name}: {e}')
-                    pbar.update(1)
-
-                feature_groups.append(fg)
-
-        return feature_groups
-
-    def _create_enhanced_tooltip_generator(self) -> TooltipGenerator:
-        """Create tooltip generator that includes related KPIs."""
-
-        class EnhancedKPITooltip(TooltipGenerator):
-            def __init__(self, kpi_visualizer):
-                self.kpi_visualizer = kpi_visualizer
-
-            def generate_tooltip(self, data_item: MapDataItem) -> str:
-                if not isinstance(data_item, KPIDataItem):
-                    return super().generate_tooltip(data_item)
-
-                kpi = data_item.kpi
-                kpi_name = kpi.get_kpi_name_with_dataset_name()
-
-                from mescal.units import Units
-                kpi_quantity = Units.get_quantity_in_pretty_unit(kpi.quantity)
-                kpi_text = Units.get_pretty_text_for_quantity(kpi_quantity, thousands_separator=' ')
-
-                html = '<table style="border-collapse: collapse;">\n'
-                html += f'  <tr><td style="padding: 4px 8px;"><strong>{kpi_name}</strong></td>' \
-                        f'<td style="text-align: right; padding: 4px 8px;">{kpi_text}</td></tr>\n'
-
-                if self.kpi_visualizer.include_related_kpis_in_tooltip and self.kpi_visualizer.study_manager:
-                    related_groups = self.kpi_visualizer.grouping_manager.get_related_kpi_groups(
-                        kpi, self.kpi_visualizer.study_manager
-                    )
-
-                    if any(not g.empty for g in related_groups.values()):
-                        for name, group in related_groups.items():
-                            if group.empty:
-                                continue
-                            html += "<tr><p>&nbsp;</p></tr>"
-                            html += f'  <tr><th colspan="2" style="text-align: left; padding: 8px;">{name}</th></tr>\n'
-                            for related_kpi in group:
-                                related_kpi_name = related_kpi.get_kpi_name_with_dataset_name()
-                                related_kpi_quantity = Units.get_quantity_in_pretty_unit(related_kpi.quantity)
-                                related_kpi_value_text = Units.get_pretty_text_for_quantity(
-                                    related_kpi_quantity,
-                                    thousands_separator=' ',
-                                )
-                                html += f'  <tr><td style="padding: 4px 8px;">{related_kpi_name}</td>' \
-                                        f'<td style="text-align: right; padding: 4px 8px;">{related_kpi_value_text}</td></tr>\n'
-
-                html += '<br><p>&nbsp;</p></table>'
-                return html
-
-        return EnhancedKPITooltip(self)
-
-    def _resolve_generator(self, generator: Union[str, FoliumObjectGenerator]) -> FoliumObjectGenerator:
-        """Resolve generator from name or return instance directly."""
-        if isinstance(generator, str):
-            if generator not in self.generators:
-                raise ValueError(f"Generator '{generator}' not registered")
-            return self.generators[generator]
-        return generator
-
-
-if __name__ == '__main__':
-    import numpy as np
-    from shapely.geometry import Point, Polygon
-
-    # Create dummy data
-    model_df = pd.DataFrame({
-        'name': ['Area1', 'Area2', 'Area3'],
-        'price': [45, 85, 150],
-        'capacity': [100, 200, 300],
-        'height': [10, 20, 15],
-        'geometry': [
-            Polygon([(0, 0), (1, 0), (1, 1), (0, 1)]),
-            Polygon([(1, 0), (2, 0), (2, 1), (1, 1)]),
-            Polygon([(0, 1), (1, 1), (1, 2), (0, 2)])
-        ]
-    })
-
-    # Create flexible style resolver with custom mappings
-    style_resolver = StyleResolver([
-        StyleMapper('color', 'price', lambda x: f'hsl({x * 2}, 70%, 50%)', str),
-        StyleMapper('width', 'capacity', lambda x: x / 100, float),
-        StyleMapper('opacity', None, 0.8, float),
-        StyleMapper('height', 'height', lambda x: x * 2, float),
-        StyleMapper('shadow_color', None, '#000000', str),
-        StyleMapper('border_color', None, 'white', str),
-    ])
-
-    # Create generators
-    area_generator = AreaGenerator(style_resolver)
-    text_generator = TextOverlayGenerator(
-        style_resolver,
-        text_formatter=lambda di: f"{di.get_styling_value('price'):.0f}€"
-    )
-
-    # Create map builder for basic usage
-    builder = FoliumMapBuilder()
-    builder.register_generator('areas', area_generator)
-
-    # Create KPI map visualizer for advanced KPI grouping (replicates old functionality)
-    kpi_style_resolver = StyleResolver([
-        StyleMapper('color', 'kpi_value', lambda x: f'hsl({x * 2}, 70%, 50%)', str),
-        StyleMapper('width', None, 2.0, float),
-        StyleMapper('opacity', None, 0.8, float),
-    ])
-
-    kpi_area_generator = AreaGenerator(kpi_style_resolver)
-    kpi_visualizer = KPIMapVisualizer(
-        generator=kpi_area_generator,
-        # study_manager=study_manager,  # Would be passed in real usage
-        include_related_kpis_in_tooltip=True,
-        kpi_attribute_category_orders={'dataset': ['base', 'scenario1', 'scenario2']},
-        kpi_attribute_keys_to_exclude_from_grouping=['name', 'object_name'],
-        kpi_attribute_sort_order=['aggregation', 'flag', 'dataset', 'value_operation']
-    )
-
-    # Create map
-    m = folium.Map(location=[0.5, 0.5], zoom_start=8)
-
-    # Add model data using builder
-    area_fg = folium.FeatureGroup(name='Areas')
-    builder.add_model_data('areas', model_df, area_fg)
-    m.add_child(area_fg)
-
-    # Add KPI data using high-level visualizer (replicates original get_feature_groups)
-    # feature_groups = kpi_visualizer.get_feature_groups(kpi_collection)
-    # for fg in feature_groups:
-    #     m.add_child(fg)
-
-    print("Enhanced map system with KPI grouping functionality! 🎉")
-    print("✅ KPIMapVisualizer replicates KPIToMapVisualizerBase.get_feature_groups()")
-    print("✅ KPIGroupingManager handles sophisticated KPI grouping")
-    print("✅ Maintains all original grouping and sorting logic")
-    print("✅ Integrates seamlessly with new generator system")
-    print("✅ Enhanced tooltips with related KPIs when enabled")
